+++ app/src/main/java/protect/card_locker/DBHelper.java
@@ -17,7 +17,7 @@ import java.nio.ByteBuffer;
 
 public class DBHelper extends SQLiteOpenHelper
 {
-    public static final String DATABASE_NAME = "LoyaltyCards.db";
+    public static final String DATABASE_NAME = "Catima.db";
     public static final int ORIGINAL_DATABASE_VERSION = 1;
     public static final int DATABASE_VERSION = 4;