+++ app/src/main/java/protect/card_locker/LoyaltyCardViewActivity.java
@@ -31,6 +31,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.zxing.BarcodeFormat;
 
 import org.json.JSONException;
@@ -44,7 +45,7 @@ import protect.card_locker.preferences.Settings;
 
 public class LoyaltyCardViewActivity extends AppCompatActivity
 {
-    private static final String TAG = "CardLocker";
+    private static final String TAG = "Catima";
     private static final double LUMINANCE_MIDPOINT = 0.5;
 
     TextView cardIdFieldView;
@@ -282,6 +283,20 @@ public class LoyaltyCardViewActivity extends AppCompatActivity
         {
             findViewById(R.id.barcode).setVisibility(View.GONE);
         }
+
+        FloatingActionButton editButton = findViewById(R.id.fabEdit);
+        editButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), LoyaltyCardEditActivity.class);
+                Bundle bundle = new Bundle();
+                bundle.putInt("id", loyaltyCardId);
+                bundle.putBoolean("update", true);
+                intent.putExtras(bundle);
+                startActivity(intent);
+                finish();
+            }
+        });
     }
 
     @Override
@@ -318,7 +333,6 @@ public class LoyaltyCardViewActivity extends AppCompatActivity
         }
 
         menu.findItem(R.id.action_share).setIcon(getIcon(R.drawable.ic_share_white, backgroundNeedsDarkIcons));
-        menu.findItem(R.id.action_edit).setIcon(getIcon(R.drawable.ic_mode_edit_white_24dp, backgroundNeedsDarkIcons));
 
         return super.onCreateOptionsMenu(menu);
     }
@@ -336,16 +350,6 @@ public class LoyaltyCardViewActivity extends AppCompatActivity
 
             case R.id.action_share:
                 importURIHelper.startShareIntent(loyaltyCard);
-                return true;
-
-            case R.id.action_edit:
-                Intent intent = new Intent(getApplicationContext(), LoyaltyCardEditActivity.class);
-                Bundle bundle = new Bundle();
-                bundle.putInt("id", loyaltyCardId);
-                bundle.putBoolean("update", true);
-                intent.putExtras(bundle);
-                startActivity(intent);
-                finish();
                 return true;
 
             case R.id.action_lock_unlock:
@@ -455,6 +459,9 @@ public class LoyaltyCardViewActivity extends AppCompatActivity
             // Move barcode to top
             barcodeImage.setScaleType(ImageView.ScaleType.FIT_START);
 
+            // Prevent centering
+            barcodeImage.setAdjustViewBounds(false);
+
             // Set current state
             barcodeIsFullscreen = true;
         }
@@ -478,6 +485,9 @@ public class LoyaltyCardViewActivity extends AppCompatActivity
 
             // Turn barcode back to normal
             barcodeImage.setLayoutParams(barcodeImageState);
+
+            // Fix barcode centering
+            barcodeImage.setAdjustViewBounds(true);
 
             // Set current state
             barcodeIsFullscreen = false;